requirement.txt
compressai==1.2.4
dm-control==1.0.13
dmc2gym==1.0.0
glfw==2.6.2
gym==0.15.7
imageio==2.9.0
imageio-ffmpeg==0.4.2
imgaug==0.2.6
matplotlib==3.7.3
mujoco==2.3.6
numpy==1.21.2
opencv-python==4.8.0.74
pandas==2.0.3
Pillow==10.0.0
PyOpenGL==3.1.7
pyOpenSSL==21.0.0
PyYAML==6.0
scikit-image==0.21.0
tensorboard==2.13.0
torch==1.11.0
torchvision==0.12.0
tqdm==4.62.3